I had to do some changes to make it work properly due to how timezones work. The system now interprets the stored database values for start_date and unlock_date as if they were already stored in the users timezone, even if they are technically stored in UTC format. This is necessary for timezones that have a UTC offset which crosses over to the next day if they start the challenge near 12:00am. This also means timezones will automatically account for movement. So if a day unlocks December 2nd, it will always be unlocked for the student even if they fly for 12 hours to a timezone -12 hours from where they were (meaning for them time has no changed, so our system should think that time has not changed for them).
It's all working and tested but we must keep in mind that technically the unlocked_dates in the DB are not true UTC values and are always interpreted based on where the user is. Carbon does automatic things when you use ISO time strings so I had to use some nasty code to make sure Carbon doesn't use UTC when it shouldn't.